# packages/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HyperDX External API",
    "description": "API for managing HyperDX alerts and dashboards",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Your HyperDX instance (http://<host>:<port>)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v2/team": {
      "get": {
        "summary": "Get Team",
        "description": "Retrieves the authenticated team's basic settings.",
        "operationId": "getTeam",
        "tags": [
          "Team"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved team"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v2/team/invitation": {
      "post": {
        "summary": "Invite Team Member",
        "description": "Creates a pending invitation for a new email address.",
        "operationId": "inviteTeamMember",
        "tags": [
          "Team"
        ],
        "responses": {
          "200": {
            "description": "Successfully created the invitation"
          },
          "400": {
            "description": "User already exists"
          }
        }
      }
    },
...
    "/api/v2/team/invitation/{id}": {
      "delete": {
        "summary": "Delete Team Invitation",
        "description": "Deletes a pending team invitation.",
        "operationId": "deleteTeamInvitation",
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the invitation"
          },
          "404": {
            "description": "Invitation not found"
          }
        }
      }
    },
    "/api/v2/team/member/{id}": {
      "delete": {
        "summary": "Remove Team Member",
        "description": "Removes a member from the team.",
        "operationId": "removeTeamMember",
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully removed the member"
          },
          "400": {
            "description": "Cannot remove yourself from the team"
          },
          "404": {
            "description": "Member not found"
          }
        }
      }
    },
...
  }
}